+++ src/app/TxService.ts
@@ -1,78 +1,21 @@
-import {
-  Constraint,
-  CreateTableMode,
-  DataType,
-  QueryClient,
-  QueryTable,
-  TableOptions,
-} from "../../deps/index.ts";
-
-import * as env from "./env.ts";
-
-export type TransactionData = {
-  txId?: number;
-  pubKey: string;
-  signature: string;
-  contractAddress: string;
-  methodId: string;
-  encodedParams: string;
-};
-
-const txOptions: TableOptions = {
-  txId: { type: DataType.Serial, constraint: Constraint.PrimaryKey },
-  pubKey: { type: DataType.VarChar, length: 66 },
-  signature: { type: DataType.VarChar, length: 64 },
-  contractAddress: { type: DataType.VarChar, length: 42 },
-  methodId: { type: DataType.VarChar, length: 10 },
-  encodedParams: { type: DataType.VarChar },
-};
+import TxStore, { TransactionData } from "./TxStore.ts";
 
 export default class TxService {
-  client: QueryClient;
-  txTable: QueryTable<TransactionData>;
-
-  private constructor(txTableName: string) {
-    this.client = new QueryClient({
-      hostname: env.PG.HOST,
-      port: env.PG.PORT,
-      user: env.PG.USER,
-      password: env.PG.PASSWORD,
-      database: env.PG.DB_NAME,
-      tls: {
-        enforce: false,
-      },
-    });
-
-    this.txTable = this.client.table<TransactionData>(txTableName);
-  }
-
-  static async create(txTableName: string): Promise<TxService> {
-    const txService = new TxService(txTableName);
-    await txService.txTable.create(txOptions, CreateTableMode.IfNotExists);
-
-    return txService;
-  }
+  constructor(public store: TxStore) {}
 
   async addTx(txData: TransactionData) {
-    await this.txTable.insert(txData);
+    await this.store.addTx(txData);
   }
 
   async txCount(): Promise<bigint> {
-    const result = await this.client.query(
-      `SELECT COUNT(*) FROM ${this.txTable.name}`,
-    );
-    return result[0].count as bigint;
+    return await this.store.txCount();
   }
 
   async getTxs(): Promise<TransactionData[]> {
-    return await this.client.query(`SELECT * FROM ${this.txTable.name}`);
+    return await this.store.getTxs();
   }
 
   async resetTable() {
-    await this.txTable.create(txOptions, CreateTableMode.DropIfExists);
-  }
-
-  async stop() {
-    await this.client.disconnect();
+    return await this.store.resetTable();
   }
 }
